对于insert，返回值为insert_id

+++ application/db_proxy/src/db_proxy.thrift
@@ -34,7 +34,7 @@ service DbProxyService
 
     // DB更新和插入
     // 返回更新或插入的记录数
-    i32 update(1: string sign, 2: i32 seq, 3: i32 update_index, 4: list<string> tokens)
+    i64 update(1: string sign, 2: i32 seq, 3: i32 update_index, 4: list<string> tokens)
 
     // update()的异步版本
     oneway void async_update(1: string sign, 2: i32 seq, 3: i32 update_index, 4: list<string> tokens)
@@ -42,11 +42,11 @@ service DbProxyService
     // 根据参数自动拼接sql，不需要模版
     // conditions Where条件结点，只有AND关系，如：(conditions[0].left op conditions[0].right) AND (conditions[1].left op conditions[1].right)
     // UPDATE tablename (tokens[0].first,tokens[1].first) VALUES (tokens[0].second,tokens[1].second) WHERE (condtion[0].left op condtion[0].right) 
-    i32 update2(1: i32 seq, 2: i32 database_index, 3: string tablename, 4: map<string, string> tokens, 5: list<Condition> conditions)
+    i64 update2(1: i32 seq, 2: i32 database_index, 3: string tablename, 4: map<string, string> tokens, 5: list<Condition> conditions)
 
     // 根据参数自动拼接sql，不需要模版
     // INSERT INTO tablename (tokens[0].first,tokens[1].first) VALUES (tokens[0].second,tokens[1].second)
-    i32 insert2(1: i32 seq, 2: i32 database_index, 3: string tablename, 4: map<string, string> tokens)
+    i64 insert2(1: i32 seq, 2: i32 database_index, 3: string tablename, 4: map<string, string> tokens)
 
     // 根据参数自动拼接sql，不需要模版
     // conditions Where条件结点，只有AND关系，如：(conditions[0].left op conditions[0].right) AND (conditions[1].left op conditions[1].right)
