+++ application/db_proxy/src/db_proxy_handler.h
@@ -29,11 +29,11 @@ public:
 
 private: // override DbProxyServiceIf
     virtual void query(DBTable& _return, const std::string& sign, const int32_t seq, const int32_t query_index, const std::vector<std::string> & tokens, const int32_t limit, const int32_t limit_start);
-    virtual int update(const std::string& sign, const int32_t seq, const int32_t update_index, const std::vector<std::string> & tokens);
+    virtual int64_t update(const std::string& sign, const int32_t seq, const int32_t update_index, const std::vector<std::string> & tokens);
     virtual void async_update(const std::string& sign, const int32_t seq, const int32_t update_index, const std::vector<std::string> & tokens);
 
-    virtual int32_t update2(const int32_t seq, const int32_t database_index, const std::string& tablename, const std::map<std::string, std::string> & tokens, const std::vector<Condition> & conditions);
-    virtual int32_t insert2(const int32_t seq, const int32_t database_index, const std::string& tablename, const std::map<std::string, std::string> & tokens);
+    virtual int64_t update2(const int32_t seq, const int32_t database_index, const std::string& tablename, const std::map<std::string, std::string> & tokens, const std::vector<Condition> & conditions);
+    virtual int64_t insert2(const int32_t seq, const int32_t database_index, const std::string& tablename, const std::map<std::string, std::string> & tokens);
     virtual void query2(DBTable& _return, const int32_t seq, const int32_t database_index, const std::string& table, const std::vector<std::string> & fields, const std::vector<Condition> & conditions, const std::string& groupby, const std::string& orderby, const int32_t limit, const int32_t limit_start);
 
 private:
@@ -42,7 +42,7 @@ private:
     void escape_tokens(void* db_connection, const std::vector<std::string>& tokens, std::vector<std::string>* escaped_tokens);
 
     // throw_exception 是否抛异常，对于同步版本需要抛，异步版本不抛
-    int do_update(bool throw_exception, const std::string& sign, const int32_t seq, const int32_t update_index, const std::vector<std::string> & tokens);
+    int64_t do_update(bool throw_exception, const std::string& sign, const int32_t seq, const int32_t update_index, const std::vector<std::string> & tokens);
 
     // 从缓存中取数据，如果取到返回true，否则返回false
     bool get_data_from_cache(DBTable& dbtable, const std::string& sql);
@@ -50,7 +50,7 @@ private:
     void add_data_to_cache(const DBTable& dbtable, const std::string& sql, int cached_seconds);
 
     // 入加SQL或写入文件中
-    int write_sql(const struct DbInfo& db_info, sys::DBConnection* db_connection, const std::string& sql);
+    int64_t write_sql(const struct DbInfo& db_info, sys::DBConnection* db_connection, const std::string& sql);
 
 private:
     typedef std::tr1::unordered_map<utils::CMd5Helper::Value, struct CachedData, utils::CMd5Helper::ValueHasher, utils::CMd5Helper::ValueComparer> CacheTable;
